Refactor embed
